refactor: separate commentMenu file
